refactor: rename types file in ingestion
