+++ src/models/raster/ingestion/types.ts
